# src/generated/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PurchaseOrder": {
				"properties": {
					"id": {
						"type": "string"
					},
					"poNumber": {
						"type": "string"
					},
					"lineItem": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time",
						"default": "2021-03-21T18:41:45.669Z"
					},
					"customerNumber": {
						"type": "string"
					},
					"paymentReference": {
						"type": "string"
					},
					"isDeleted": {
						"type": "boolean",
						"default": false
					}
				},
				"required": [
					"id",
					"poNumber",
					"lineItem",
					"quantity",
					"customerNumber",
					"paymentReference"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_PurchaseOrder.poNumber-or-lineItem-or-quantity-or-customerNumber-or-paymentReference_": {
				"properties": {
					"poNumber": {
						"type": "string"
					},
					"lineItem": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"customerNumber": {
						"type": "string"
					},
					"paymentReference": {
						"type": "string"
					}
				},
				"required": [
					"poNumber",
					"lineItem",
					"quantity",
					"customerNumber",
					"paymentReference"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"POCreationParams": {
				"$ref": "#/components/schemas/Pick_PurchaseOrder.poNumber-or-lineItem-or-quantity-or-customerNumber-or-paymentReference_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "purchase-tsoa",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Paul Fremantle ",
			"email": "pzfreo@github.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/purchase/{uuid}": {
			"get": {
				"operationId": "GetPurchase",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PurchaseOrder"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePO",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/PurchaseOrder"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/POCreationParams"
							}
						}
					}
				}
			}
		},
		"/purchase": {
			"get": {
				"operationId": "GetAllPurchases",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"href": {
												"type": "string"
											}
										},
										"required": [
											"href"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PurchaseOrder"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/POCreationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}